Extract FullFeatureListView to a separate file
